requirements.txt
pdf2image==1.16.3
Pillow==10.2.0
anthropic==0.34.2
python-dotenv==1.0.0
pypdf>=3.0.0
reportlab>=4.0.0
